requirements.txt
playwright
psycopg2-binary
screeninfo
openpyxl
pandas